+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - Fix the issue that the parser takes too long to parse some code (#57).
 - Fix the issue that the code `let x= 42; {...}` could not be parsed (since the parser expected a space between `x` and `=`).
 - Fix the issue that compiling some code causes a compiler crash (#59).
+- Fix the issue that `FFI_CALL[() f(CLongLong)]` (or some of other similar expressions) could not be parsed (#62).
 
 ### Fixed